requirements.txt
alembic==1.8.1
click==8.1.3
Flask==2.2.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.2
MarkupSafe==2.1.1
paho-mqtt==1.6.1
psycopg2-binary==2.9.3
pyngrok==5.1.0
python-dotenv==0.21.0
pyuwsgi==2.0.20
PyYAML==6.0
SQLAlchemy==1.4.41
Werkzeug==2.2.2
zipp==3.8.1
